requirements.txt
pytesseract==0.3.10
Pillow==10.0.0
lxml==4.9.3 